// src/rmmz/system/core/booleanOptions/gameSetting/types.ts
export interface System_BooleanGameOptions {
  optAutosave: boolean;
  optDisplayTp: boolean;
  optDrawTitle: boolean;
  optExtraExp: boolean;
  optFloorDeath: boolean;
  optFollowers: boolean;
  optKeyItemsNumber: boolean;
  optSideView: boolean;
  optSlipDeath: boolean;
  optTransparent: boolean;
  optSplashScreen?: boolean;
  optMessageSkip: boolean;
}

export type SystemLable_BooleanOptionsRMMMZ = Record<
  keyof System_BooleanGameOptions,
  string
>;